Added .vimrc to the repo
I am using Linux and it works just fine for me. It might also work on MacOS as well. To use it on Windows  PC you need to check your environment/home variable otherwise it will be tricky for you to do these all in Windows.
This is the .vimrc file that you need to create if you don't have. Just import the setting I have if you want to. Otherwise detailed description/comment is also given, read and choose the functionality/plugins/settings you want to add.